[REFACT] : unified modal structure - button 바로 밑 형제 자리에 모달 컴포넌트 배치
